Removed unused variable for issue #143
Removed unused loan_to_value variable. Verified that all calculation
are still accurate by using https://usmortgagecalculator.org/ . The
variable is no longer used because it's only purpose was to check if a
user has supplied a down payment of greater than or equal to 20%. that
check is now being handled as a pop up window on the front end.